Refactor: 채팅 헤더부분 폴더화
