+++ src/components/chat/characterHeader/FriendShip.tsx
